Update messy to 6.0.0 and use its JSON features instead of upgrading/downgrading the body before matching.
This removes a great deal of complexity and makes assertions on message bodies a bit more predictable.

It is a breaking change because it removes support for implicitly "downcasting" by satisfying the body against a "lower fidelity" type, eg. a JSON body against a string, or a text body against a Buffer. I never saw a good use case for that anyway.